added collected coin number to level finished and game lost dialogs. tweaked the ratio between shooterLevel and shooting speed and the ratio between shooterLevel and min and max ball hp

+++ app/src/main/java/com/example/fivepointsproject/Wave.java
@@ -5,7 +5,7 @@ public class Wave {
     public double minHp;
     public double maxHp;
 
-    public Wave(int numBalls, int minHp, int maxHp) {
+    public Wave(int numBalls, double minHp, double maxHp) {
         this.numBalls = numBalls;
         this.minHp = minHp;
         this.maxHp = maxHp;
